// Frontend/my-app/src/common/define-type.ts
export interface CompartmentData{
    ID_khoangrac: string;
    NhanRac: string;
    KhoiLuong: number;
    AnhRac: string[]

}

export interface BinData{
    ID_thungrac: string;
    ViTriThungRac: string;
    Khoangrac: CompartmentData[];
}

export interface GarbageQuantity{
    KhoiLuong: number,
    NhanRac: string
}
export interface BarChartData{
    Ngay: string;
    Rac: GarbageQuantity[]
}

export interface GetImageResponse{
    ID_khoangrac: string;
    AnhRac: string[];
    NhanRac: string;
}